+++ modules/model_flux.py
@@ -225,7 +225,8 @@ def load_flux(checkpoint_info, diffusers_load_config): # triggered by opts.sd_ch
 
     if shared.opts.teacache_enabled:
         from modules import teacache
-        diffusers.FluxTransformer2DModel.forward = teacache.teacache_forward
+        shared.log.debug(f'Transformers cache: type=teacache patch=forward cls={diffusers.FluxTransformer2DModel.__name__}')
+        diffusers.FluxTransformer2DModel.forward = teacache.teacache_flux_forward
 
     # load overrides if any
     if shared.opts.sd_unet != 'Default':